+++ src/components/app/app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './components/layout/app-header/Header';
-import Main from './components/layout/main/Main';
+import Header from '../layout/app-header/header';
+import Main from '../layout/main/main';
 
 function App() {
   return (
